+++ backend/__init__.py
@@ -1,0 +1,1 @@
+"""Backend package for core modules (data collection, querying, monitoring)."""
